Implementar PUT, GET, DELETE de Usuarios

+++ src/main/java/com/aluracursos/challenge_forohub/controller/UsuarioController.java
@@ -3,12 +3,13 @@ package com.aluracursos.challenge_forohub.controller;
 
 import com.aluracursos.challenge_forohub.domain.perfil.Perfil;
 import com.aluracursos.challenge_forohub.domain.perfil.PerfilRepository;
-import com.aluracursos.challenge_forohub.domain.usuario.DatosRegistroUsuario;
-import com.aluracursos.challenge_forohub.domain.usuario.DatosRespuestaUsuario;
-import com.aluracursos.challenge_forohub.domain.usuario.Usuario;
-import com.aluracursos.challenge_forohub.domain.usuario.UsuarioRepository;
+import com.aluracursos.challenge_forohub.domain.usuario.*;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,6 +18,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/usuarios")
@@ -32,8 +34,27 @@ public class UsuarioController {
     private PasswordEncoder passwordEncoder;
 
     @GetMapping
-    public List<Usuario> listarUsuarios() {
-        return usuarioRepository.findAll();
+    public ResponseEntity<Page<DatosListadoUsuario>> listadoUsuarios(@PageableDefault(page = 0, size = 10, sort = {"nombre"}) Pageable pageable) {
+        return ResponseEntity.ok(usuarioRepository.findAll(pageable).map(DatosListadoUsuario::new));
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<DatosRespuestaUsuario> retornaDatosUsuarios(@PathVariable Long id) {
+        // Verificación de que el ID es positivo
+        if (id == null || id <= 0) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(null);  // Devuelve 400 si el ID es nulo o negativo
+        }
+
+        return usuarioRepository.findById(id)
+                .map(usuario -> ResponseEntity.ok(new DatosRespuestaUsuario(usuario)))
+                .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+    }
+
+    @GetMapping("/buscar")
+    public Page<DatosListadoUsuario> buscarPorNombreOCorreoElectronico(@RequestParam(required = false) String nombre, @RequestParam(required = false) String correo, @PageableDefault(size = 10, sort = "nombre") Pageable pageable) {
+        return usuarioRepository.buscarPorNombreOCorreoElectronico(nombre, correo, pageable)
+                .map(DatosListadoUsuario::new);
     }
 
     @PostMapping
@@ -61,17 +82,45 @@ public class UsuarioController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Usuario> actualizarUsuario(@PathVariable Long id, @RequestBody Usuario usuario) {//TODO usar DTO
-        return usuarioRepository.findById(id).map(u -> {
-            //usar DTO
-//            u.setNombre(usuario.getNombre());
-//            u.setCorreoElectronico(usuario.getCorreoElectronico());
-            return ResponseEntity.ok(usuarioRepository.save(u));
-        }).orElse(ResponseEntity.notFound().build());
+    @Transactional
+    public ResponseEntity<DatosRespuestaUsuario> actualizarUsuario(@PathVariable Long id, @RequestBody @Valid DatosActualizarUsuario datosActualizarUsuario) {
+        // Validar si el ID es válido
+        if (id == null || id <= 0) {
+            throw new IllegalArgumentException("El id proporcionado no es válido: " + id);
+        }
+
+        // Buscar el usuario con Optional
+        var usuarioOptional = usuarioRepository.findById(id);
+        if (!usuarioOptional.isPresent()) {
+            throw new IllegalArgumentException("El usuario con ID " + id + " no existe.");
+        }
+
+        Usuario usuario = usuarioOptional.get();
+
+        // Encriptar constaseña usando BCrypt solo si está presente
+        String passwordEncriptado = null;
+        if (datosActualizarUsuario.contrasena() != null) {
+            passwordEncriptado = passwordEncoder.encode(datosActualizarUsuario.contrasena());
+        }
+
+        // Buscar perfiles en la BD
+        List<Perfil> nuevosPerfiles = perfilRepository.findByNombreIn(datosActualizarUsuario.perfiles());
+
+        usuario.actualizarDatosUsuario(datosActualizarUsuario, nuevosPerfiles, passwordEncriptado);
+        return ResponseEntity.ok(new DatosRespuestaUsuario(usuario));
     }
 
     @DeleteMapping("/{id}")
+    @Transactional
     public ResponseEntity<Usuario> eliminarUsuario(@PathVariable Long id) {
+        if (id == null || id <= 0) {
+            throw new IllegalArgumentException("El id " + id + " no existe.");
+        }
+        Optional<Usuario> optionalUsuario = usuarioRepository.findById(id);
+        if (!optionalUsuario.isPresent()) {
+            throw new IllegalArgumentException("El usuario con id " + id + " no existe.");
+        }
+
         usuarioRepository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
